requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
datasets>=2.14.0
huggingface-hub>=0.16.0
numpy<2.0

pytorch-lightning>=2.2.0
transformers>=4.40.0

datasets>=2.14.0
pyyaml>=6.0.0

tensorboard>=2.15.0
